requirements.txt
openai
requests
TTS
ffmpeg-python
python-slugify
python-dotenv